refactor: Repository 인터페이스와 JDBC 구현체 분리
