+++ src/main/java/roomescape/dao/ReservationTimeRepository.java
@@ -1,59 +1,15 @@
 package roomescape.dao;
 
-import java.time.LocalTime;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
 import roomescape.domain.ReservationTime;
 
-@Repository
-public class ReservationTimeRepository {
+public interface ReservationTimeRepository {
 
-    private final JdbcTemplate jdbcTemplate;
-    private final SimpleJdbcInsert simpleJdbcInsert;
+    List<ReservationTime> findAll();
 
-    public ReservationTimeRepository(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.simpleJdbcInsert = new SimpleJdbcInsert(jdbcTemplate)
-                .withTableName("reservation_time")
-                .usingGeneratedKeyColumns("id");
-    }
+    ReservationTime findById(Long id);
 
-    public List<ReservationTime> findAll() {
-        String sql = "SELECT id, start_at FROM reservation_time";
+    ReservationTime save(ReservationTime reservationTime);
 
-        return jdbcTemplate.query(sql, (rs, rowNum) -> new ReservationTime(
-                rs.getLong("id"),
-                LocalTime.parse(rs.getString("start_at"))
-        ));
-    }
-
-    public ReservationTime findById(Long id) {
-        String sql = "SELECT id, start_at FROM reservation_time WHERE id = ?";
-
-        return jdbcTemplate.queryForObject(sql, (rs, ronNum) -> new ReservationTime(
-                rs.getLong("id"),
-                LocalTime.parse(rs.getString("start_at"))
-        ), id);
-    }
-
-    public ReservationTime save(ReservationTime reservationTime) {
-        Map<String, Object> parameters = new HashMap<>();
-        parameters.put("start_at", reservationTime.getStartAt().toString());
-
-        Number key = simpleJdbcInsert.executeAndReturnKey(parameters);
-
-        return new ReservationTime(
-                key.longValue(),
-                reservationTime.getStartAt()
-        );
-    }
-
-    public void deleteById(Long id) {
-        String sql = "DELETE FROM reservation_time WHERE id = ?";
-        jdbcTemplate.update(sql, id);
-    }
+    void deleteById(Long id);
 }
